// App/client/types/post.ts
export type PostInfoState = {
    author: {
        displayName: string;
        email: string;
        photoURL: string;
        uid: string;
        userNum: number;
        __v: number;
        _id: string;
    };
    content: string;
    createdAt: string;
    image: string;
    postNum: number;
    replyNum: number;
    title: string;
    updatedAt: string;
    __v: number;
    _id: string;
};